requirements.txt
fastapi==0.68.0
uvicorn==0.12.2
fastapi_sqlalchemy
python-dotenv